% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_loader.R
\name{UC_load}
\alias{UC_load}
\title{Load an .RData file from Unity Catalog}
\usage{
UC_load(uc_path, envir = parent.frame(), ...)
}
\arguments{
\item{uc_path}{UC path to the .RData/.rda file.}

\item{envir}{Environment to load objects into.}

\item{...}{Passed to internal downloader (e.g. \code{verbose}, \code{retry_on_auth}).}
}
\value{
(Invisible) character vector of object names created by \code{load()}.
}
\description{
Downloads the file to a temporary location, loads it into \code{envir}, then deletes the temp file.
Requires \code{DATABRICKS_HOST} to be set (e.g. via cluster init script).
}
