% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucLoader-package.R
\docType{package}
\name{ucfs}
\alias{ucLoader}
\alias{ucLoader-package}
\alias{ucfs}
\title{Minimal UC file helpers (REST)}
\description{
Convenience helpers to list, download, upload and load files stored in
Databricks Unity Catalog Volumes via the REST API. Requires
\code{DATABRICKS_HOST} and \code{DATABRICKS_TOKEN}.
}
\section{Key functions}{

\itemize{
\item \code{\link[=connect]{connect()}}
\item \code{\link[=UC_list_files]{UC_list_files()}}
\item \code{\link[=download]{download()}}
\item \code{\link[=upload]{upload()}}
\item \code{\link[=UC_load]{UC_load()}}
\item \code{\link[=UC_read.csv]{UC_read.csv()}}
\item \code{\link[=UC_save]{UC_save()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ridout-m/ucLoader}
  \item Report bugs at \url{https://github.com/ridout-m/ucLoader/issues}
}

}
\author{
\strong{Maintainer}: Mackenzie Ridout \email{Mackenzie.Ridout@defra.gov.uk}

}
\keyword{internal}
